+++ src/mpl_wrap/__init__.py
@@ -4,3 +4,4 @@ from importlib import metadata
 __version__ = metadata.version(__name__)
 
 from mpl_wrap.wrap import *
+from mpl_wrap.axes_wrap import *